Refactor routes
