+++ src/main/kotlin/ru/neexol/routes/apiRoute.kt
@@ -1,0 +1,10 @@
+package ru.neexol.routes
+
+import io.ktor.server.routing.*
+import ru.neexol.routes.api.*
+
+fun Route.apiRoute() {
+    updateRoute()
+    checksumRoute()
+    scheduleRoute()
+}
